debezium/dbz#197 Refactor ByLogicalTableRouter to extend ToLogicalTopicRouter and update metadata references
