+++ debezium-connect-plugins/src/main/resources/META-INF/services/org.apache.kafka.connect.transforms.Transformation
@@ -1,5 +1,6 @@
 io.debezium.transforms.ByLogicalTableRouter
 io.debezium.transforms.EnforceRecordSize
+io.debezium.transforms.ToLogicalTopicRouter
 io.debezium.transforms.ExtractChangedRecordState
 io.debezium.transforms.ExtractNewRecordState
 io.debezium.transforms.ExtractSchemaToNewRecord
